+++ tests/test.py
@@ -128,11 +128,10 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
         if not cmd:
             cmd = ['true']
         return {"Tty": False, "Image": img, "Cmd": cmd,
-                "AttachStdin": False, "Memory": 0,
+                "AttachStdin": False,
                 "AttachStderr": True, "AttachStdout": True,
                 "StdinOnce": False,
                 "OpenStdin": False, "NetworkDisabled": False,
-                "MemorySwap": 0
                 }
 
     def test_ctor(self):
@@ -341,11 +340,10 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
         self.assertEqual(json.loads(args[1]['data']),
                          json.loads('''
                             {"Tty": false, "Image": "busybox", "Cmd": ["true"],
-                             "AttachStdin": false, "Memory": 0,
+                             "AttachStdin": false,
                              "AttachStderr": true, "AttachStdout": true,
                              "StdinOnce": false,
-                             "OpenStdin": false, "NetworkDisabled": false,
-                             "MemorySwap": 0}'''))
+                             "OpenStdin": false, "NetworkDisabled": false}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -365,12 +363,11 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["ls", "/mnt"], "AttachStdin": false,
-                             "Volumes": {"/mnt": {}}, "Memory": 0,
+                             "Volumes": {"/mnt": {}},
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
-                             "NetworkDisabled": false,
-                             "MemorySwap": 0}'''))
+                             "NetworkDisabled": false}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -390,12 +387,11 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["ls", "/mnt"], "AttachStdin": false,
-                             "Volumes": {"/mnt": {}}, "Memory": 0,
+                             "Volumes": {"/mnt": {}},
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
-                             "NetworkDisabled": false,
-                             "MemorySwap": 0}'''))
+                             "NetworkDisabled": false}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -413,7 +409,7 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["ls"], "AttachStdin": false,
-                             "Memory": 0, "ExposedPorts": {
+                             "ExposedPorts": {
                                 "1111/tcp": {},
                                 "2222/udp": {},
                                 "3333/tcp": {}
@@ -421,8 +417,7 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
-                             "NetworkDisabled": false,
-                             "MemorySwap": 0}'''))
+                             "NetworkDisabled": false}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -440,13 +435,11 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["hello"], "AttachStdin": false,
-                             "Memory": 0,
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
                              "NetworkDisabled": false,
-                             "Entrypoint": "cowsay",
-                             "MemorySwap": 0}'''))
+                             "Entrypoint": "cowsay"}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -464,13 +457,11 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["ls"], "AttachStdin": false,
-                             "Memory": 0,
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
                              "NetworkDisabled": false,
-                             "CpuShares": 5,
-                             "MemorySwap": 0}'''))
+                             "CpuShares": 5}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -488,14 +479,12 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["ls"], "AttachStdin": false,
-                             "Memory": 0,
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
                              "NetworkDisabled": false,
                              "Cpuset": "0,1",
-                             "CpusetCpus": "0,1",
-                             "MemorySwap": 0}'''))
+                             "CpusetCpus": "0,1"}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -513,13 +502,11 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
                          json.loads('''
                             {"Tty": false, "Image": "busybox",
                              "Cmd": ["ls"], "AttachStdin": false,
-                             "Memory": 0,
                              "AttachStderr": true,
                              "AttachStdout": true, "OpenStdin": false,
                              "StdinOnce": false,
                              "NetworkDisabled": false,
-                             "WorkingDir": "/root",
-                             "MemorySwap": 0}'''))
+                             "WorkingDir": "/root"}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -535,11 +522,10 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
         self.assertEqual(json.loads(args[1]['data']),
                          json.loads('''
                             {"Tty": false, "Image": "busybox", "Cmd": ["true"],
-                             "AttachStdin": true, "Memory": 0,
+                             "AttachStdin": true,
                              "AttachStderr": true, "AttachStdout": true,
                              "StdinOnce": true,
-                             "OpenStdin": true, "NetworkDisabled": false,
-                             "MemorySwap": 0}'''))
+                             "OpenStdin": true, "NetworkDisabled": false}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
 
@@ -585,78 +571,95 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
         self.assertEqual(json.loads(args[1]['data']),
                          json.loads('''
                             {"Tty": false, "Image": "busybox", "Cmd": ["true"],
-                             "AttachStdin": false, "Memory": 0,
+                             "AttachStdin": false,
                              "AttachStderr": true, "AttachStdout": true,
                              "StdinOnce": false,
-                             "OpenStdin": false, "NetworkDisabled": false,
-                             "MemorySwap": 0}'''))
+                             "OpenStdin": false, "NetworkDisabled": false}'''))
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
         self.assertEqual(args[1]['params'], {'name': 'marisa-kirisame'})
 
     def test_create_container_with_mem_limit_as_int(self):
         try:
-            self.client.create_container('busybox', 'true',
-                                         mem_limit=128.0)
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    mem_limit=128.0
+                )
+            )
         except Exception as e:
             self.fail('Command should not raise exception: {0}'.format(e))
 
         args = fake_request.call_args
         data = json.loads(args[1]['data'])
-        self.assertEqual(data['Memory'], 128.0)
+        self.assertEqual(data['HostConfig']['Memory'], 128.0)
 
     def test_create_container_with_mem_limit_as_string(self):
         try:
-            self.client.create_container('busybox', 'true',
-                                         mem_limit='128')
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    mem_limit='128'
+                )
+            )
         except Exception as e:
             self.fail('Command should not raise exception: {0}'.format(e))
 
         args = fake_request.call_args
         data = json.loads(args[1]['data'])
-        self.assertEqual(data['Memory'], 128.0)
+        self.assertEqual(data['HostConfig']['Memory'], 128.0)
 
     def test_create_container_with_mem_limit_as_string_with_k_unit(self):
         try:
-            self.client.create_container('busybox', 'true',
-                                         mem_limit='128k')
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    mem_limit='128k'
+                )
+            )
         except Exception as e:
             self.fail('Command should not raise exception: {0}'.format(e))
 
         args = fake_request.call_args
         data = json.loads(args[1]['data'])
-        self.assertEqual(data['Memory'], 128.0 * 1024)
+        self.assertEqual(data['HostConfig']['Memory'], 128.0 * 1024)
 
     def test_create_container_with_mem_limit_as_string_with_m_unit(self):
         try:
-            self.client.create_container('busybox', 'true',
-                                         mem_limit='128m')
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    mem_limit='128m'
+                )
+            )
+
         except Exception as e:
             self.fail('Command should not raise exception: {0}'.format(e))
 
         args = fake_request.call_args
         data = json.loads(args[1]['data'])
-        self.assertEqual(data['Memory'], 128.0 * 1024 * 1024)
+        self.assertEqual(data['HostConfig']['Memory'], 128.0 * 1024 * 1024)
 
     def test_create_container_with_mem_limit_as_string_with_g_unit(self):
         try:
-            self.client.create_container('busybox', 'true',
-                                         mem_limit='128g')
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    mem_limit='128g'
+                )
+            )
         except Exception as e:
             self.fail('Command should not raise exception: {0}'.format(e))
 
         args = fake_request.call_args
         data = json.loads(args[1]['data'])
-        self.assertEqual(data['Memory'], 128.0 * 1024 * 1024 * 1024)
+        self.assertEqual(
+            data['HostConfig']['Memory'], 128.0 * 1024 * 1024 * 1024
+        )
 
     def test_create_container_with_mem_limit_as_string_with_wrong_value(self):
-        self.assertRaises(docker.errors.DockerException,
-                          self.client.create_container,
-                          'busybox', 'true', mem_limit='128p')
+        self.assertRaises(
+            docker.errors.DockerException, create_host_config, mem_limit='128p'
+        )
 
-        self.assertRaises(docker.errors.DockerException,
-                          self.client.create_container,
-                          'busybox', 'true', mem_limit='1f28')
+        self.assertRaises(
+            docker.errors.DockerException, create_host_config, mem_limit='1f28'
+        )
 
     def test_start_container(self):
         try:
@@ -804,6 +807,83 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
         expected_payload = self.base_create_payload()
         expected_payload['HostConfig'] = create_host_config()
         expected_payload['HostConfig']['Binds'] = ["/tmp:/mnt:rw"]
+        self.assertEqual(json.loads(args[1]['data']), expected_payload)
+        self.assertEqual(args[1]['headers'],
+                         {'Content-Type': 'application/json'})
+        self.assertEqual(
+            args[1]['timeout'],
+            DEFAULT_TIMEOUT_SECONDS
+        )
+
+    def test_create_container_with_binds_mode(self):
+        try:
+            mount_dest = '/mnt'
+            mount_origin = '/tmp'
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    binds={mount_origin: {
+                        "bind": mount_dest,
+                        "mode": "z",
+                    }}
+                )
+            )
+        except Exception as e:
+            self.fail('Command should not raise exception: {0}'.format(e))
+
+        args = fake_request.call_args
+        self.assertEqual(args[0][0], url_prefix +
+                         'containers/create')
+        expected_payload = self.base_create_payload()
+        expected_payload['HostConfig'] = create_host_config()
+        expected_payload['HostConfig']['Binds'] = ["/tmp:/mnt:z"]
+        self.assertEqual(json.loads(args[1]['data']), expected_payload)
+        self.assertEqual(args[1]['headers'],
+                         {'Content-Type': 'application/json'})
+        self.assertEqual(
+            args[1]['timeout'],
+            DEFAULT_TIMEOUT_SECONDS
+        )
+
+    def test_create_container_with_binds_mode_and_ro_error(self):
+        try:
+            mount_dest = '/mnt'
+            mount_origin = '/tmp'
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    binds={mount_origin: {
+                        "bind": mount_dest,
+                        "mode": "z",
+                        "ro": True,
+                    }}
+                )
+            )
+        except ValueError:
+            return
+
+        self.fail('Command should raise ValueError')
+
+    def test_create_container_with_binds_list(self):
+        try:
+            self.client.create_container(
+                'busybox', 'true', host_config=create_host_config(
+                    binds=[
+                        "/tmp:/mnt/1:ro",
+                        "/tmp:/mnt/2",
+                    ],
+                )
+            )
+        except Exception as e:
+            self.fail('Command should not raise exception: {0}'.format(e))
+
+        args = fake_request.call_args
+        self.assertEqual(args[0][0], url_prefix +
+                         'containers/create')
+        expected_payload = self.base_create_payload()
+        expected_payload['HostConfig'] = create_host_config()
+        expected_payload['HostConfig']['Binds'] = [
+            "/tmp:/mnt/1:ro",
+            "/tmp:/mnt/2",
+        ]
         self.assertEqual(json.loads(args[1]['data']), expected_payload)
         self.assertEqual(args[1]['headers'],
                          {'Content-Type': 'application/json'})
@@ -1354,6 +1434,37 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
             args[1]['timeout'], DEFAULT_TIMEOUT_SECONDS
         )
 
+    def test_create_container_with_named_volume(self):
+        try:
+            mount_dest = '/mnt'
+            volume_name = 'name'
+            self.client.create_container(
+                'busybox', 'true',
+                host_config=create_host_config(
+                    binds={volume_name: {
+                        "bind": mount_dest,
+                        "ro": False
+                    }}),
+                volume_driver='foodriver',
+            )
+        except Exception as e:
+            self.fail('Command should not raise exception: {0}'.format(e))
+
+        args = fake_request.call_args
+        self.assertEqual(args[0][0], url_prefix +
+                         'containers/create')
+        expected_payload = self.base_create_payload()
+        expected_payload['VolumeDriver'] = 'foodriver'
+        expected_payload['HostConfig'] = create_host_config()
+        expected_payload['HostConfig']['Binds'] = ["name:/mnt:rw"]
+        self.assertEqual(json.loads(args[1]['data']), expected_payload)
+        self.assertEqual(args[1]['headers'],
+                         {'Content-Type': 'application/json'})
+        self.assertEqual(
+            args[1]['timeout'],
+            DEFAULT_TIMEOUT_SECONDS
+        )
+
     def test_resize_container(self):
         try:
             self.client.resize(
@@ -1799,15 +1910,16 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
             timeout=DEFAULT_TIMEOUT_SECONDS
         )
 
-    def test_inspect_container_empty_id(self):
-        try:
-            self.client.inspect_container('')
-        except docker.errors.NullResource as e:
-            self.assertEqual(
-                e.args[0], 'image or container param is undefined'
-            )
-        else:
-            self.fail('Command expected NullResource exception')
+    def test_inspect_container_undefined_id(self):
+        for arg in None, '', {True: True}:
+            try:
+                self.client.inspect_container(arg)
+            except docker.errors.NullResource as e:
+                self.assertEqual(
+                    e.args[0], 'image or container param is undefined'
+                )
+            else:
+                self.fail('Command expected NullResource exception')
 
     def test_container_stats(self):
         try:
@@ -1980,15 +2092,16 @@ class DockerClientTest(Cleanup, base.BaseTestCase):
             timeout=DEFAULT_TIMEOUT_SECONDS
         )
 
-    def test_inspect_image_empty_id(self):
-        try:
-            self.client.inspect_image('')
-        except docker.errors.NullResource as e:
-            self.assertEqual(
-                e.args[0], 'image or container param is undefined'
-            )
-        else:
-            self.fail('Command expected NullResource exception')
+    def test_inspect_image_undefined_id(self):
+        for arg in None, '', {True: True}:
+            try:
+                self.client.inspect_image(arg)
+            except docker.errors.NullResource as e:
+                self.assertEqual(
+                    e.args[0], 'image or container param is undefined'
+                )
+            else:
+                self.fail('Command expected NullResource exception')
 
     def test_insert_image(self):
         try: